refactor: move alloy extensions to new alloy_ext crate
