+++ integration-tests/Cargo.toml
@@ -13,6 +13,7 @@ build = "build.rs"
 
 [dependencies]
 zksync_os_server.workspace = true
+zksync_os_alloy_ext.workspace = true
 zksync_os_contract_interface.workspace = true
 zksync_os_tx_validators.workspace = true
 zksync_os_types.workspace = true
