+++ integration-tests/src/prover_tester.rs
@@ -1,6 +1,6 @@
-use crate::dyn_wallet_provider::EthDynProvider;
-use crate::network::Zksync;
-use crate::provider::ZksyncApi;
+use zksync_os_alloy_ext::dyn_wallet_provider::EthDynProvider;
+use zksync_os_alloy_ext::network::Zksync;
+use zksync_os_alloy_ext::provider::ZksyncApi;
 use alloy::eips::BlockNumberOrTag;
 use alloy::primitives::{U256, keccak256};
 use alloy::providers::{DynProvider, Provider};
