+++ integration-tests/src/upgrade/tester.rs
@@ -6,8 +6,9 @@ use super::interfaces;
 use crate::Tester;
 use crate::assert_traits::ReceiptAssert;
 use crate::config::load_chain_config;
-use crate::dyn_wallet_provider::EthDynProvider;
-use crate::provider::{ZksyncApi as _, ZksyncTestingProvider as _};
+use crate::provider::ZksyncTestingProvider as _;
+use zksync_os_alloy_ext::dyn_wallet_provider::EthDynProvider;
+use zksync_os_alloy_ext::provider::ZksyncApi as _;
 use crate::upgrade::interfaces::ChainAssetHandlerBase::ChainAssetHandlerBaseInstance;
 use crate::upgrade::interfaces::ChainTypeManagerV30::ChainTypeManagerV30Instance;
 use crate::upgrade::interfaces::FacetCut;
