+++ integration-tests/tests/protocol/interop.rs
@@ -13,9 +13,10 @@ use anyhow::{Context, Result};
 use zksync_os_contract_interface::Bridgehub;
 use zksync_os_contract_interface::IMailbox::NewPriorityRequest;
 use zksync_os_integration_tests::assert_traits::ProviderAssert;
-use zksync_os_integration_tests::dyn_wallet_provider::EthDynProvider;
+use zksync_os_alloy_ext::dyn_wallet_provider::EthDynProvider;
+use zksync_os_alloy_ext::provider::ZksyncApi;
 use zksync_os_integration_tests::{
-    GatewayTester, Tester, assert_traits::ReceiptAssert, contracts::TestERC20, provider::ZksyncApi,
+    GatewayTester, Tester, assert_traits::ReceiptAssert, contracts::TestERC20,
 };
 use zksync_os_rpc_api::types::LogProofTarget;
 use zksync_os_types::{
