+++ integration-tests/tests/rpc/deployment_filter.rs
@@ -8,7 +8,7 @@ use anyhow::Result;
 use tokio::time::{Duration, timeout};
 use zksync_os_integration_tests::assert_traits::ReceiptAssert;
 use zksync_os_integration_tests::contracts::EventEmitter;
-use zksync_os_integration_tests::dyn_wallet_provider::EthWalletProvider;
+use zksync_os_alloy_ext::dyn_wallet_provider::EthWalletProvider;
 use zksync_os_integration_tests::{DeploymentFilterConfig, GatewayTester};
 
 /// The default rich wallet address (derived from the well-known test private key).
