+++ integration-tests/tests/rpc/filter.rs
@@ -8,7 +8,7 @@ use alloy::sol_types::SolEvent;
 use zksync_os_integration_tests::assert_traits::ReceiptAssert;
 use zksync_os_integration_tests::contracts::EventEmitter;
 use zksync_os_integration_tests::contracts::EventEmitter::{EventEmitterInstance, TestEvent};
-use zksync_os_integration_tests::dyn_wallet_provider::EthDynProvider;
+use zksync_os_alloy_ext::dyn_wallet_provider::EthDynProvider;
 use zksync_os_integration_tests::{CURRENT_TO_L1, Tester, test_multisetup};
 
 trait FilterSuite: Sized {
