+++ integration-tests/tests/rpc/pubsub.rs
@@ -11,7 +11,7 @@ use tokio::time::error::Elapsed;
 use zksync_os_integration_tests::assert_traits::ReceiptAssert;
 use zksync_os_integration_tests::contracts::EventEmitter;
 use zksync_os_integration_tests::contracts::EventEmitter::{EventEmitterInstance, TestEvent};
-use zksync_os_integration_tests::dyn_wallet_provider::EthDynProvider;
+use zksync_os_alloy_ext::dyn_wallet_provider::EthDynProvider;
 use zksync_os_integration_tests::{CURRENT_TO_L1, Tester, test_multisetup};
 
 trait PubsubSuite: Sized {
