+++ lib/alloy_ext/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "zksync_os_alloy_ext"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+zksync_os_contract_interface.workspace = true
+zksync_os_rpc_api.workspace = true
+zksync_os_types.workspace = true
+
+alloy = { workspace = true, default-features = false, features = ["json-rpc", "rpc-types"] }
+async-trait.workspace = true
+anyhow.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
